Code cleanup Move classes and struct into their own separate files
