+++ App/Sources/Core/Models/Commands/Bundled/MoveToWorkspaceCommand.swift
@@ -1,0 +1,4 @@
+struct MoveToWorkspaceCommand: Identifiable, Hashable, Codable {
+  let id: String
+  let workspace: WorkspaceCommand
+}
